Move BaseLayerType to types folder
