+++ src/interface/src/app/map/map.component.spec.ts
@@ -1,4 +1,3 @@
-import { Region } from './../types/region.types';
 import { HarnessLoader } from '@angular/cdk/testing';
 import { TestbedHarnessEnvironment } from '@angular/cdk/testing/testbed';
 import { NO_ERRORS_SCHEMA } from '@angular/core';
@@ -8,12 +7,13 @@ import { MatCheckboxModule } from '@angular/material/checkbox';
 import { MatCheckboxHarness } from '@angular/material/checkbox/testing';
 import { MatRadioModule } from '@angular/material/radio';
 import { MatRadioGroupHarness } from '@angular/material/radio/testing';
-import { of, BehaviorSubject } from 'rxjs';
+import { BehaviorSubject, of } from 'rxjs';
 
 import { MapService } from '../map.service';
 import { PopupService } from '../popup.service';
-import { BaseLayerType, MapComponent } from './map.component';
 import { SessionService } from './../session.service';
+import { BaseLayerType, Region } from './../types';
+import { MapComponent } from './map.component';
 
 describe('MapComponent', () => {
   let component: MapComponent;
@@ -111,12 +111,14 @@ describe('MapComponent', () => {
 
     // Act: select the terrain base layer
     await radioButtonGroup.checkRadioButton({ label: 'Terrain' });
+
     // Assert: expect that the map contains the terrain base layer
     expect(component.changeBaseLayer).toHaveBeenCalled();
     expect(component.map.hasLayer(MapComponent.hillshade_tiles));
 
     // Act: select the road base layer
     await radioButtonGroup.checkRadioButton({ label: 'Road' });
+
     // Assert: expect that the map contains the road base layer
     expect(component.changeBaseLayer).toHaveBeenCalled();
     expect(component.map.hasLayer(MapComponent.open_street_maps_tiles));
@@ -129,12 +131,14 @@ describe('MapComponent', () => {
 
     // Act: uncheck the HUC-12 checkbox
     await checkbox.uncheck();
+
     // Assert: expect that the map does not contain the HUC-12 layer
     expect(component.toggleHUC12BoundariesLayer).toHaveBeenCalled();
     expect(component.map.hasLayer(component.HUC12BoundariesLayer)).toBeFalse();
 
     // Act: check the HUC-12 checkbox
     await checkbox.check();
+
     // Assert: expect that the map contains the HUC-12 layer
     expect(component.toggleHUC12BoundariesLayer).toHaveBeenCalled();
     expect(component.map.hasLayer(component.HUC12BoundariesLayer)).toBeTrue();
@@ -143,14 +147,18 @@ describe('MapComponent', () => {
   it('should toggle existing projects layer', async () => {
     component.ngAfterViewInit();
     spyOn(component, 'toggleExistingProjectsLayer').and.callThrough();
+
     // Act: uncheck the existing projects checkbox
     const checkbox = await loader.getHarness(MatCheckboxHarness.with({ name: 'existing-projects-toggle' }));
     await checkbox.uncheck();
+
     // Assert: expect that the map removes the existing projects layer
     expect(component.toggleExistingProjectsLayer).toHaveBeenCalled();
     expect(component.map.hasLayer(component.existingProjectsLayer)).toBeFalse();
+
     // Act: check the existing projects checkbox
     await checkbox.check();
+
     // Assert: expect that the map adds the existing projects layer
     expect(component.toggleExistingProjectsLayer).toHaveBeenCalled();
     expect(component.map.hasLayer(component.existingProjectsLayer)).toBeTrue();
