+++ src/interface/src/app/types/index.ts
@@ -1,1 +1,2 @@
+export * from "./layer.types";
 export * from "./region.types";
