+++ src/interface/src/app/types/layer.types.ts
@@ -1,0 +1,4 @@
+export enum BaseLayerType {
+  Road,
+  Terrain,
+}
